api naming convention adjusted to follow api design specifications
